authentication + authorization

+++ models/campground.js
@@ -9,6 +9,10 @@ const CampgroundSchema = new Schema({
     image: String,
     description: String,
     location: String,
+    author: {
+        type: Schema.Types.ObjectId,
+        ref: 'User'
+    },
     reviews: [
         {
             type: Schema.Types.ObjectId,
@@ -19,7 +23,7 @@ const CampgroundSchema = new Schema({
 
 //Query middleware
 CampgroundSchema.post('findOneAndDelete', async function(campground){
-    if(campground.reviews.length){
+    if(campground.reviews && campground.reviews.length){
         await Review.deleteMany({
             _id: {$in: campground.reviews}
         })
